venv-requirements.txt
snowflake-snowpark-python[pandas]
apache-airflow-providers-snowflake
psycopg2-binary